pyproject.toml
[tool.poetry]
name = "ai-research-assistant-local-rag-agent-ollama"
version = "0.1.0"
description = ""
authors = ["kamran945 <kamran945@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
feedparser = "^6.0.11"
pandas = "^2.2.3"
requests = "^2.32.3"
langchain = "^0.3.7"
langchain-community = "^0.3.5"
pypdf = "^5.1.0"
pinecone = "^5.3.1"
langchain-huggingface = "^0.1.2"
langchain-groq = "^0.2.1"
tavily-python = "^0.5.0"
langgraph = "^0.2.45"
semantic-router = "^0.0.72"
streamlit = "^1.40.0"
langchain-core = "^0.3.15"
python-dotenv = "^1.0.1"
xmltodict = "^0.14.2"
langchain-pinecone = "^0.2.0"
langchain-ollama = "^0.2.0"
arxiv = "^2.1.3"
pymupdf = "^1.24.14"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
